+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Supports Typo3 8.7.0 - 8.7.99
 
+- [FEATURE] Implement memory function for qualifications of the employees.
 - [UPDATE] Fix status designation.
 
 ## 2.3.1